requirements.txt
Django>=5.2
djangorestframework>=3.14.0